# Reduce enqueue and dequeue allocations