--- a/bigqueue.go
+++ b/bigqueue.go
@@ -24,6 +24,8 @@ var (
 
 // MmapQueue implements Queue interface.
 type MmapQueue struct {
+	bytesReader
+	stringReader
 	conf      *bqConfig
 	am        *arenaManager
 	md        *metadata