@moredure 2021-04-19
Before
```
cpu: Intel(R) Core(TM) i5-7600 CPU @ 3.50GHz
BenchmarkEnqueue
BenchmarkEnqueue/ArenaSize-4KB/MessageSize-128B/MaxMem-12KB
BenchmarkEnqueue/ArenaSize-4KB/MessageSize-128B/MaxMem-12KB-4         	  225560	      6584 ns/op	      45 B/op	       1 allocs/op
BenchmarkEnqueue/ArenaSize-4KB/MessageSize-128B/MaxMem-40KB
BenchmarkEnqueue/ArenaSize-4KB/MessageSize-128B/MaxMem-40KB-4         	  351213	      5216 ns/op	      45 B/op	       1 allocs/op
```
After
```go
pu: Intel(R) Core(TM) i5-7600 CPU @ 3.50GHz
BenchmarkEnqueue
BenchmarkEnqueue/ArenaSize-4KB/MessageSize-128B/MaxMem-12KB
BenchmarkEnqueue/ArenaSize-4KB/MessageSize-128B/MaxMem-12KB-4         	  325824	      3537 ns/op	      18 B/op	       0 allocs/op
BenchmarkEnqueue/ArenaSize-4KB/MessageSize-128B/MaxMem-40KB
BenchmarkEnqueue/ArenaSize-4KB/MessageSize-128B/MaxMem-40KB-4         	  320850	      3857 ns/op	      18 B/op	       0 allocs/op
```